@@ -1,20 +1,20 @@
 name: status_alert
 description: Display Apple system-like self-hiding status alerts. It is well suited for notifying user without interrupting user flow.
-version: 0.1.3
+version: 1.0.0-nullsafety.1
 homepage: https://github.com/yako-dev/flutter-status-alert
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  flare_flutter: ^2.0.6
+  flare_flutter: ^3.0.0-nullsafety.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  pedantic: ^1.9.2
+  pedantic: ^1.11.0
 
 flutter:
   fonts:
