@@ -8,27 +8,27 @@ import 'package:status_alert/src/models/status_alert_text_configuration.dart';
 import 'package:status_alert/src/utils/colors.dart';
 
 class StatusAlertBaseWidget extends StatefulWidget {
-  final PopupMediaConfiguration configuration;
+  final PopupMediaConfiguration? configuration;
   final BorderRadius borderRadius;
-  final Color backgroundColor;
+  final Color? backgroundColor;
   final Alignment alignment;
   final EdgeInsets padding;
   final EdgeInsets margin;
-  final String title;
-  final String subtitle;
-  final Duration duration;
-  final StatusAlertTextConfiguration titleOptions;
-  final StatusAlertTextConfiguration subtitleOptions;
-  final VoidCallback onHide;
-  final double blurPower;
+  final String? title;
+  final String? subtitle;
+  final Duration? duration;
+  final StatusAlertTextConfiguration? titleOptions;
+  final StatusAlertTextConfiguration? subtitleOptions;
+  final VoidCallback? onHide;
+  final double? blurPower;
 
   const StatusAlertBaseWidget({
-    Key key,
-    @required this.margin,
-    @required this.padding,
-    @required this.alignment,
-    @required this.borderRadius,
-    @required this.backgroundColor,
+    Key? key,
+    required this.margin,
+    required this.padding,
+    required this.alignment,
+    required this.borderRadius,
+    required this.backgroundColor,
     this.title,
     this.onHide,
     this.subtitle,
@@ -46,10 +46,10 @@ class StatusAlertBaseWidget extends StatefulWidget {
 class __TDBaseWidgetState extends State<StatusAlertBaseWidget>
     with SingleTickerProviderStateMixin {
   Duration animationDuration = Duration(milliseconds: 200);
-  AnimationController scaleController;
-  AnimationController animationController;
-  Animation<double> scaleAnimation;
-  Animation<double> fadeAnimation;
+  AnimationController? scaleController;
+  AnimationController? animationController;
+  late Animation<double> scaleAnimation;
+  late Animation<double> fadeAnimation;
 
   final Tween<double> scaleTween = Tween<double>(begin: 0.9, end: 1.0);
   final Tween<double> fadeTween = Tween<double>(begin: 0.0, end: 1.0);
@@ -65,23 +65,23 @@ class __TDBaseWidgetState extends State<StatusAlertBaseWidget>
       duration: Duration(milliseconds: 200),
       vsync: this,
     );
-    scaleAnimation = scaleTween.animate(animationController);
-    fadeAnimation = fadeTween.animate(animationController);
+    scaleAnimation = scaleTween.animate(animationController!);
+    fadeAnimation = fadeTween.animate(animationController!);
     if (mounted) {
       animationController?.forward();
       await Future.delayed(animationDuration);
-      await Future.delayed(widget.duration);
+      await Future.delayed(widget.duration!);
     }
     if (mounted) {
       animationController?.reverse();
       await Future.delayed(animationDuration);
     }
-    widget.onHide();
+    widget.onHide!();
   }
 
   @override
   void dispose() {
-    animationController.dispose();
+    animationController!.dispose();
     super.dispose();
   }
 
@@ -124,28 +124,25 @@ class __TDBaseWidgetState extends State<StatusAlertBaseWidget>
 
     List<Widget> content = <Widget>[];
     if (widget.configuration is WidgetConfiguration) {
-      WidgetConfiguration config = widget.configuration;
+      WidgetConfiguration config = widget.configuration as WidgetConfiguration;
       content.add(config.widget);
     } else {
       if (widget.configuration is IconConfiguration) {
-        IconConfiguration config = widget.configuration;
+        IconConfiguration config = widget.configuration as IconConfiguration;
         content.add(Padding(
           padding: config.margin,
           child: Icon(
             config.icon,
             key: config.key,
             size: config.size ?? screenWidth * 0.35,
-            color:
-                config.color ?? Theme.of(context).brightness == Brightness.light
-                    ? lightAccent
-                    : darkAccent,
+            color: config.color ?? getThemeColor(),
             semanticLabel: config.semanticLabel,
             textDirection: config.textDirection,
           ),
         ));
       }
       if (widget.configuration is FlareConfiguration) {
-        FlareConfiguration config = widget.configuration;
+        FlareConfiguration config = widget.configuration as FlareConfiguration;
         content.add(Padding(
           padding: config.margin,
           child: SizedBox(
@@ -173,24 +170,21 @@ class __TDBaseWidgetState extends State<StatusAlertBaseWidget>
         content.add(Padding(
           padding: const EdgeInsets.only(top: 4.0),
           child: Text(
-            widget.title,
-            key: widget.titleOptions.key,
-            style: widget.titleOptions.style.copyWith(
-              color: widget.titleOptions.style.color ??
-                      Theme.of(context).brightness == Brightness.light
-                  ? lightAccent
-                  : darkAccent,
+            widget.title!,
+            key: widget.titleOptions!.key,
+            style: widget.titleOptions!.style.copyWith(
+              color: widget.titleOptions!.style.color ?? getThemeColor(),
             ),
-            locale: widget.titleOptions.locale,
-            softWrap: widget.titleOptions.softWrap,
-            maxLines: widget.titleOptions.maxLines,
-            overflow: widget.titleOptions.overflow,
-            textAlign: widget.titleOptions.textAlign,
-            strutStyle: widget.titleOptions.strutStyle,
-            textDirection: widget.titleOptions.textDirection,
-            textWidthBasis: widget.titleOptions.textWidthBasis,
-            semanticsLabel: widget.titleOptions.semanticsLabel,
-            textScaleFactor: widget.titleOptions.textScaleFactor,
+            locale: widget.titleOptions!.locale,
+            softWrap: widget.titleOptions!.softWrap,
+            maxLines: widget.titleOptions!.maxLines,
+            overflow: widget.titleOptions!.overflow,
+            textAlign: widget.titleOptions!.textAlign,
+            strutStyle: widget.titleOptions!.strutStyle,
+            textDirection: widget.titleOptions!.textDirection,
+            textWidthBasis: widget.titleOptions!.textWidthBasis,
+            semanticsLabel: widget.titleOptions!.semanticsLabel,
+            textScaleFactor: widget.titleOptions!.textScaleFactor,
           ),
         ));
       }
@@ -198,23 +192,23 @@ class __TDBaseWidgetState extends State<StatusAlertBaseWidget>
         content.add(Padding(
           padding: const EdgeInsets.only(top: 4.0),
           child: Text(
-            widget.subtitle,
-            key: widget.subtitleOptions.key,
-            style: widget.subtitleOptions.style.copyWith(
-                color: widget.subtitleOptions.style.color ??
+            widget.subtitle!,
+            key: widget.subtitleOptions!.key,
+            style: widget.subtitleOptions!.style.copyWith(
+                color: widget.subtitleOptions!.style.color ??
                     (Theme.of(context).brightness == Brightness.light
                         ? lightAccent
                         : darkAccent)),
-            locale: widget.subtitleOptions.locale,
-            softWrap: widget.subtitleOptions.softWrap,
-            maxLines: widget.subtitleOptions.maxLines,
-            overflow: widget.subtitleOptions.overflow,
-            textAlign: widget.subtitleOptions.textAlign,
-            strutStyle: widget.subtitleOptions.strutStyle,
-            textDirection: widget.subtitleOptions.textDirection,
-            textWidthBasis: widget.subtitleOptions.textWidthBasis,
-            semanticsLabel: widget.subtitleOptions.semanticsLabel,
-            textScaleFactor: widget.subtitleOptions.textScaleFactor,
+            locale: widget.subtitleOptions!.locale,
+            softWrap: widget.subtitleOptions!.softWrap,
+            maxLines: widget.subtitleOptions!.maxLines,
+            overflow: widget.subtitleOptions!.overflow,
+            textAlign: widget.subtitleOptions!.textAlign,
+            strutStyle: widget.subtitleOptions!.strutStyle,
+            textDirection: widget.subtitleOptions!.textDirection,
+            textWidthBasis: widget.subtitleOptions!.textWidthBasis,
+            semanticsLabel: widget.subtitleOptions!.semanticsLabel,
+            textScaleFactor: widget.subtitleOptions!.textScaleFactor,
           ),
         ));
       }
@@ -248,5 +242,11 @@ class __TDBaseWidgetState extends State<StatusAlertBaseWidget>
         ),
       ),
     );
+  }
+
+  Color getThemeColor() {
+    return Theme.of(context).brightness == Brightness.light
+        ? lightAccent
+        : darkAccent;
   }
 }
